Show different music page based on music driver setting
