+++ src/components/assets/MusicViewer.css
@@ -1,6 +1,6 @@
 .MusicViewer {
   /* background: red; */
-  /* position: fixed; */
+  position: fixed;
   right: 300px;
   left: 0px;
   top: 38px;
